+++ packages/terra-clinical-header/src/terra-dev-site/test/clinical-header/NoTitleRightContentHeader.test.jsx
@@ -1,13 +1,12 @@
 import React from 'react';
 
 import Header from '../../../Header';
+import styles from './styles.module.scss';
 
 const content = (
   <div
     id="headerTest--content"
-    style={{
-      backgroundColor: 'black', height: '30px', width: '300px', margin: '0 10px 0 0',
-    }}
+    className={styles.startcontent}
   />
 );