+++ packages/terra-clinical-item-display/src/terra-dev-site/test/clinical-item-display/display/ScalingIconItemDisplay.test.jsx
@@ -2,12 +2,13 @@ import React from 'react';
 import IconSuccess from 'terra-icon/lib/icon/IconSuccess';
 
 import ItemDisplay from '../../../../ItemDisplay';
+import styles from './styles.module.scss';
 
 const views = () => (
   <div>
     <ItemDisplay text="Test Text" icon={<IconSuccess />} id="Default" />
-    <ItemDisplay text="Test Text" icon={<IconSuccess />} id="Larger" style={{ fontSize: '30px' }} />
-    <ItemDisplay text="Test Text" icon={<IconSuccess />} id="Smaller" style={{ fontSize: '10px' }} />
+    <ItemDisplay text="Test Text" icon={<IconSuccess />} id="Larger" className={styles.large} />
+    <ItemDisplay text="Test Text" icon={<IconSuccess />} id="Smaller" className={styles.smaller} />
   </div>
 );