+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Replaced Object.assign syntax with Object spread syntax
+* Add comments for ESlint 6
 
 2.9.0 - (August 14, 2019)
 ----------