+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Replaced Object.assign syntax with Object spread syntax
+* Add comments for ESlint 6
 
 3.7.0 - (August 14, 2019)
 ----------